refactor: scaffold for CPU/GPU sort backends
