+++ Cargo.toml
@@ -15,13 +15,20 @@ default-run = "viewer"
 
 
 [features]
-default = ["io_flexbuffers", "io_ply"]
+default = [
+  "io_flexbuffers",
+  "io_ply",
+  "sort_radix",
+]
+
+debug_gpu = []
 
 io_bincode2 = ["bincode2", "flate2"]
 io_flexbuffers = ["flexbuffers"]
 io_ply = ["ply-rs"]
 
-debug_gpu = []
+sort_radix = []
+sort_rayon = ["rayon"]
 
 
 [dependencies]
@@ -34,7 +41,9 @@ flate2  = { version = "1.0", optional = true }
 flexbuffers = { version = "2.0", optional = true }
 ply-rs = { version = "0.1", optional = true }
 rand = "0.8"
+rayon = { version = "1.8.0", optional = true }
 serde = "1.0"
+static_assertions = "1.1.0"
 wgpu = "0.17.1"
